extract castling
