+++ src/state.rs
@@ -1,35 +1,8 @@
 use crate::{
-    board::Board, from_to_step::FromToStep, m0ve::Move, piece::Piece, piece::Piece::*,
-    player::Player, player::Player::*, pos::Pos,
+    board::Board, castling::Castling, from_to_step::FromToStep, m0ve::Move, piece::Piece,
+    piece::Piece::*, player::Player, player::Player::*, pos::Pos,
 };
 use itertools::Itertools;
-
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub struct CastleAbility {
-    pub king: bool,
-    pub queen: bool,
-}
-
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub struct Castling {
-    pub white: CastleAbility,
-    pub black: CastleAbility,
-}
-
-impl Castling {
-    pub fn initial() -> Self {
-        Castling {
-            white: CastleAbility {
-                king: true,
-                queen: true,
-            },
-            black: CastleAbility {
-                king: true,
-                queen: true,
-            },
-        }
-    }
-}
 
 #[derive(Debug, Clone)]
 pub struct State {
